+++ CHANGES.md
@@ -1,6 +1,10 @@
 1.2.3
 =====
 
+- Fix a regression in cloudpickle and python3.8 causing an error when trying to
+  pickle property objects.
+  ([PR #329](https://github.com/cloudpipe/cloudpickle/pull/329)).
+
 - Fix a bug when a thread imports a module while cloudpickle iterates
   over the module list
   ([PR #322](https://github.com/cloudpipe/cloudpickle/pull/322)).